\name{SimLongiMMM2}
\alias{SimLongiMMM2}
\title{
X
}
\description{
X
}
\usage{
SimLongiMMM2(data, response, group, time, id, covariates = NULL, contrasts = NULL, type = "Dunnett", base = 1, alternative = "two.sided", level = 0.95, refdist = "normal")
}
\arguments{
  \item{data}{x}
  \item{response}{x}
  \item{group}{x}
  \item{time}{x}
  \item{id}{x}
  \item{covariates}{x}
  \item{contrasts}{x}
  \item{type}{x}
  \item{base}{x}
  \item{alternative}{x}
  \item{level}{x}
  \item{refdist}{x}
}
\details{
xxx
}
\value{
xxx
}
\references{
xxx
}
\author{
xxx
}
\note{
xxx
}
\seealso{
xxx
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
}
\keyword{htest}