\name{SimLongi-package}
\alias{SimLongi-package}
\docType{package}
\title{
A package for XXX
}
\description{
More about what it does (maybe more than one line)
}
\details{
\tabular{ll}{
Package: \tab SimLongi\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-12-02\cr
License: \tab What license is it under?\cr
}
% An overview of how to use the package, including the most important functions
}
\author{
Philip Pallmann <pallmann@biostat.uni-hannover.de>
}
\references{
% Literature or other references for background information
}
% Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory
\keyword{ package }
\seealso{
% Optional links to other man pages, e.g. 
% \code{\link[<pkg>:<pkg>-package]{<pkg>}}
}
\examples{
## SimLongi()
## SimLongiMMM()
}
